feat(qt-generator): optional PARAMETERS on every Qt surface; refuse `-> ?T`
`?T` was expressible in LIDL and handled by the C++ and Rust generators, but the
Qt generators had zero optional-aware code — `qt-generator/` did not contain the
string "optional". An optional parameter therefore did not compile: the Qt
surface spelled it QVariant while the author's impl took std::optional<T>.

Parameters now work on all four Qt surfaces:

  provider   `?T` decodes through the CANONICAL codec —
             logos::fromJson<std::optional<T>>(qvariantToNlohmann(p), "p") —
             so an absent or null slot is the empty optional and anything else
             is decoded as T by the SAME decoder a required T would get.
             Optionality widens the domain; it does not switch type checking off.
  events     an empty optional marshals to the invalid QVariant. Without this,
             std::optional<T> hit QVariant::fromValue and put an unregistered
             metatype on the wire.
  consumer   optional record fields, honouring BOTH spellings (`? n: T` and
             `n: ?T`) via fieldIsOptional() so they cannot drift apart. An empty
             field omits its key rather than writing null.
  metadata   an additive per-parameter "optional" flag in getMethods().

An optional RETURN is REFUSED, on the qt and consumer backends alike, with the
reason in the message. An empty `?T` is JSON null, and null is already how this
path reports a failed call: logos_json_convert maps it to an invalid QVariant,
which core_service reports as METHOD_FAILED. "Found nothing" would be
indistinguishable from "the call failed" for every non-Rust caller. Supporting
it means NARROWING that rule to methods whose published metadata declares an
optional return — never deleting it, which would silently convert every
pre-existing failure into a success.

Two pre-existing bugs fixed in passing (see goldens/README.md, which classifies
every line of golden churn):

  * `args.at(i)` was an unchecked QList read past the end whenever a caller sent
    fewer arguments than the contract declares. Now `args.value(i)`, plus an
    explicit arity gate that answers dispatch_failed with a count.
  * `int`/`uint` returns were `static_cast<int>` — but they are int64_t/uint64_t,
    so every value past 2^31 was silently truncated on the way back to Qt.

+++ tests/qt-generator/goldens/plain_qt/plain_module_qt_glue.h
@@ -95,12 +95,12 @@ public:
 
     qlonglong echo_int(qlonglong n) {
         auto _result = m_impl.echo_int(static_cast<int64_t>(n));
-        return static_cast<int>(_result);
+        return static_cast<qlonglong>(_result);
     }
 
     qulonglong echo_uint(qulonglong n) {
         auto _result = m_impl.echo_uint(static_cast<uint64_t>(n));
-        return static_cast<int>(_result);
+        return static_cast<qulonglong>(_result);
     }
 
     bool echo_bool(bool b) {
